requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
oauthlib==3.1.1
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tweepy==4.4.0
urllib3==1.26.8
